requirements.txt
pytest == 7.2.1
pylint == 2.16.0
pycryptodome == 3.17.0
flask == 2.2.3
psycopg2
pyjwt